+++ canaille/core/mails.py
@@ -70,7 +70,7 @@ def send_password_reset_mail(user, mail):
 
     return send_email(
         subject=subject,
-        recipient=user.preferred_email,
+        recipient=mail,
         text=text_body,
         html=html_body,
         attachements=[(logo_cid, logo_filename, logo_raw)] if logo_filename else None,
@@ -111,7 +111,7 @@ def send_password_initialization_mail(user, email):
 
     return send_email(
         subject=subject,
-        recipient=user.preferred_email,
+        recipient=email,
         text=text_body,
         html=html_body,
         attachements=[(logo_cid, logo_filename, logo_raw)] if logo_filename else None,